# 06-service-recommender/mcp-211/mcp-211-search-v2-service/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: support@211.org
    name: Support 211 National Data Platform
    url: https://register.211.org/home/contact
  description: "Returns service-at-location search results from the 211 National Data\
    \ Platform's resource repository search index and values that can be used to filter\
    \ search results.\n\nA common use case for this API is to return a list of abbeviated\
    \ service at location resources (i.e. search results) using the '/keyword' endpoint,\
    \ then use values from within the search results to retrieve the details for a\
    \ selected service-at-location resource using the 'query/service-at-location-details'\
    \ endpoint on the Query V2 API.\n\nThe Search V2 API organizes operations into\
    \ two categories:\n\n* Search: returns a list of abbreviated service-at-location\
    \ resources (i.e. search results) from a GET or POST request that can be configured\
    \ to filter results using operation parameters, such as tags, service areas, taxonomy\
    \ terms, etc. The Search operations can also be configured to return facets for\
    \ targeting specific resources in downstream searches, selecting fields used for\
    \ search, and setting the type of geographic search filter (e.g. filter results\
    \ using the geographic boundary of a location or distance between the target and\
    \ the service location). The Search operation supports two methods: GET and POST.\
...
    \ for more details.\n\n"
...
servers:
...
security:
- apiKeyHeader: []
- apiKeyQuery: []
paths:
  /filters/service-area-values:
    get:
      description: "Returns a list of service area values for a given geography type\
        \ (e.g. county). Type values are: locality, postalCode, county, state, country.\r\
        \n\r\nA service area is the geographic area that a service is available. Returned\
        \ values can be used in the service areas filter on the Search/Keyword API\
        \ operation (POST method only).\r\n\r\nEnter the 'scope' value to filter values\
        \ to the geographic area type (eg. county). Enter the 'dataOwners' value(s)\
        \ to filter values on the entered 211 data owner(s) (eg. 211ventura or 211ventura,211northcarolina).\
        \ \r\nRefer to the DataOwners filter for a list of 211 data owner(s) that\
        \ can be used. \r\n\r\nUse the typeFilter and typeFilterValue to filter a\
        \ dataOwner(s) service area values on one or more higher geographical types.\
        \ As an example, 211 North Carolina has service areas outside the state of\
        \ North Carolina assigned to several of its resources. Without a filter, service\
        \ areas outside \r\nof North Carolina will be returned. Remove values for\
        \ service areas outside of North Carolina by setting the 'typeFilter' to 'state'\
        \ and 'typeFilterValues' to 'North Carolina'. Set multiple values for typeFilterValues\
        \ by separating values with a comma (e.g. North Carolina,California).\r\n\r\
        \nExample: enter 'county' as 'type' and '211ventura' as 'dataOwners' to return\
        \ a complete list of all service area county values. Use the results to create\
        \ a search filter that can be added to the search filters \r\nin the the Search/Keyword\
        \ API operation (POST method) to filter service at location resources (eg.\
        \ {\"field\": \"serviceAreas\", \"value\": [{\"type\": \"county\",\"value\"\
        : \"Ventura\"},{\"type\": \"county\",\"value\": \"San Luis Obispo\"}]} )."
      operationId: get-filters-service-area-values-type-type-dataowners-dataowners
      parameters:
      - description: "geography type to filter service area values (eg. locality,\
          \ postalCode, county, state, country). Multiple type values not supported."
        explode: true
        in: query
        name: type
        required: true
        schema:
          type: string
        style: form
      - description: "211 data owner(s) to filter service area values (e.g. 211ventura\
          \ OR 211ventura,211northcarolina). See DataOwners API fitler operation for\
          \ 211 center values."
        explode: true
        in: query
        name: dataOwners
        required: true
        schema:
          type: string
        style: form
      - description: geography type to filter type values (e.g. state). Multiple types
          not supported.
        explode: true
        in: query
        name: typeFilter
        required: false
        schema:
          type: string
        style: form
      - description: "value(s) used by type filtergeography type to filter location\
          \ values (eg. North Carolina OR North Carolina, California). Multiple values\
          \ supported."
        explode: true
        in: query
        name: typeFilterValues
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
              - string
              schema:
                items:
                  type: string
                type: array
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: "Filter: Service Area Geographies"
      tags:
      - Filters
  /filters/location-values:
    get:
      description: "Returns a list of distinct location values for a given geography\
        \ type (e.g. county). A location is where a service is delivered. Possible\
        \ values are: locality, postalCode, county, state, country.\r\n\r\nReturned\
        \ values can be used in the locationAddresses parameter on the Search/Keyword\
        \ API operation (POST method method). \r\n\r\nExample: enter 'county' as 'type'\
        \ and '211ventura' as 'dataOwners' to return a complete list of all counties\
        \ assigned to 211ventura locations. Use the typeFilter and typeFilterValue\
        \ to filter a dataOwner(s) \r\nlocation values on one or more higher geographical\
        \ types. As an example, 211 North Carolina has locations outside the state\
        \ of North Carolina assigned to several of its resources. Without a filter,\
        \ locations outside \r\nof North Carolina will be returned. Remove values\
        \ for locations outside of North Carolina by setting the 'typeFilter' to 'state'\
        \ and 'typeFilterValues' to 'NC'. Set multiple values for typeFilterValues\
        \ by separating values with a comma (e.g. NC,CA).\r\n\r\nUse the results to\
        \ create a search filter for the Search/Keyword API operation (POST method)\
        \ (eg. {\"field\": \"address/county\", \"value\": [{\"type\": \"county\",\"\
        value\": \"Ventura\"},{\"type\": \"county\",\"value\": \"San Luis Obispo\"\
        }]} )."
      operationId: get-filters-location-values-type-type-dataowners-dataowners
      parameters:
      - description: "geography type to filter location values (eg. locality, postalCode,\
          \ county, state, country). Multiple types not supported."
        explode: true
        in: query
        name: type
        required: true
        schema:
          type: string
        style: form
      - description: "data owner(s) to filter value (e.g. 211ventura OR 211ventura,211northcarolina).\
          \ See DataOwners API fitler operation for 211 center values."
        explode: true
        in: query
        name: dataOwners
        required: true
        schema:
          type: string
        style: form
      - description: geography type to filter type values (e.g. state). Multiple types
          not supported.
        explode: true
        in: query
        name: typeFilter
        required: false
        schema:
          type: string
        style: form
      - description: "value(s) used by type filtergeography type to filter location\
          \ values (eg. North Carolina OR North Carolina, California). Multiple values\
          \ supported."
        explode: true
        in: query
        name: typeFilterValues
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
              - string
              schema:
                items:
                  type: string
                type: array
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: "Filter: Location Geographies"
      tags:
      - Filters
  /filters/data-owners:
    get:
      description: "Returns all 211 center data owners and data owners display name\
        \ values. A 211 data owner is the creator or steward of the resource data.\
        \ \r\nReturned values can be used in the 'dataOwners' parameter of the Search/Keyword\
        \ API operation (GET and POST methods).\r\n\r\nExample 1: Use the results\
        \ to create a search filter that can be added to the search filters in the\
        \ of the Search/Keyword API operation (POST method) to filter service at location\
        \ resources (eg. {\"field\": \"dataOwner\", \"value\": [\"211ventura\",\"\
        211northcarolina\"]} ).\r\n\r\nExample 2: Use the results to create a search\
        \ filter that can be added to the search filters in the of the Search/Keyword\
        \ API operation (POST method) to filter service at location resources (eg.\
        \ {\"field\": \"dataOwnerDisplayName\", \"value\": [\"211 Ventura\",\"211\
        \ North Carolina\"]} )."
      operationId: get-filters-data-owners
      responses:
        "200":
          content:
            application/json:
              example:
              - dataOwner: string
                dataOwnerDisplayName: string
              schema:
                items:
                  $ref: "#/components/schemas/DataOwnerDto"
                type: array
          description: OK
        "404":
          description: Not Found
      summary: "Filter: Data Owners"
      tags:
      - Filters
  /filters/taxonomy-level2-terms:
    get:
      description: "Returns all level 2 taxonomy term values, grouped by level 1 term.\
        \ This API operation can be used to create a list of topics and subtopics.\r\
        \n\r\nList can be filtered on one or more 211 data owners"
      operationId: get-filters-taxonomy-level2-terms
      parameters:
      - description: "211 data owner(s) to filter values (e.g. 211ventura or 211ventura,211northcarolina).\
          \ \r\n            See DataOwner API fitler operation for 211 center values.\
          \ Leave blank if wanting all level 2 terms for all 211s."
        explode: true
        in: query
        name: dataOwners
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    $ref: "#/components/schemas/StringStringValueTuple"
                  type: array
                type: object
          description: OK
        "404":
          description: Not Found
      summary: "Filter: Taxonomy Level 2 Terms with Level 1 Parent"
      tags:
      - Filters
...
components:
  schemas:
...
    DataOwnerDto:
      additionalProperties: false
      example:
        dataOwnerDisplayName: dataOwnerDisplayName
        dataOwner: dataOwner
      properties:
        dataOwner:
          nullable: true
          title: dataOwner
          type: string
        dataOwnerDisplayName:
          nullable: true
          title: dataOwnerDisplayName
          type: string
      title: DataOwnerDto
      type: object
...
    StringStringValueTuple:
      additionalProperties: false
      title: StringStringValueTuple
      type: object
...
  securitySchemes:
    apiKeyHeader:
      in: header
      name: Api-Key
      type: apiKey
    apiKeyQuery:
      in: query
      name: api-key
      type: apiKey